Changes made according to comments
