+++ src/main/scala/com/activegrid/models/JsonSupport.scala
@@ -11,9 +11,9 @@ import spray.json.{DefaultJsonProtocol, DeserializationException, JsString, JsVa
 trait JsonSupport extends DefaultJsonProtocol with SprayJsonSupport {
   val logger = LoggerFactory.getLogger(getClass)
   implicit val appSettingsFormat = jsonFormat(AppSettings.apply, "id", "settings", "authSettings")
-  implicit val siteFormat = jsonFormat(Site.apply,"id","siteName","groupBy")
+  implicit val siteFormat = jsonFormat(Site.apply, "id", "siteName", "groupBy")
 
-  implicit object apmProviderFormat extends RootJsonFormat[APMProvider]{
+  implicit object apmProviderFormat extends RootJsonFormat[APMProvider] {
     override def write(obj: APMProvider): JsValue = {
       logger.info(s"Writing APMProvider json : ${obj.provider.toString}")
       JsString(obj.provider.toString)
@@ -23,14 +23,12 @@ trait JsonSupport extends DefaultJsonProtocol with SprayJsonSupport {
       logger.info(s"Reading json value : ${json.toString}")
       json match {
         case JsString(str) => APMProvider.toProvider(str)
-        case _=> throw new DeserializationException("Unable to deserialize the Provider data")
+        case _ => throw new DeserializationException("Unable to deserialize the Provider data")
       }
     }
   }
 
-  implicit val apmServerDetailsFormat = jsonFormat(APMServerDetails.apply , "id","name","serverUrl","monitoredSite","provider","headers")
-
-
+  implicit val apmServerDetailsFormat = jsonFormat(APMServerDetails.apply, "id", "name", "serverUrl", "monitoredSite", "provider", "headers")
 
 
 }
