+++ src/main/scala/com/activegrid/models/Neo4jRep.scala
@@ -6,7 +6,7 @@ import org.neo4j.graphdb.Node
   * Created by nagulmeeras on 30/09/16.
   */
 trait Neo4jRep[T <: BaseEntity] {
-  def toNeo4jGraph(): Node = ???
+  def toNeo4jGraph(): Node
 
-  def fromNeo4jGraph(nodeId: Long): T = ???
+  def fromNeo4jGraph(nodeId: Long): T
 }
